test(billing): 💍 split dfu flash tests into per-case files and add resolveFlashStatus unit test
